[BE] feat: 프로젝트 어노테이션 조회를 다시 간단하게 분리
