+++ backend/src/main/java/site/pathos/domain/subProject/controller/SubProjectController.java
@@ -14,16 +14,4 @@ import site.pathos.domain.subProject.service.SubProjectService;
 @RequestMapping("/api/subprojects")
 @RequiredArgsConstructor
 public class SubProjectController {
-
-    private final SubProjectService subProjectService;
-
-    @Operation(summary = "서브 프로젝트 상세 조회", description = "서브 프로젝트 ID로 해당 서브 프로젝트의 상세 정보를 조회합니다.")
-    @GetMapping("/{subProjectId}")
-    public ResponseEntity<SubProjectResponseDto> getSubProject(
-            @Parameter(description = "조회할 서브 프로젝트의 ID", example = "1")
-            @PathVariable("subProjectId") Long subProjectId
-    ) {
-        SubProjectResponseDto response = subProjectService.getSubProject(subProjectId);
-        return ResponseEntity.ok(response);
-    }
 }
