+++ backend/src/main/java/site/pathos/domain/subProject/service/SubProjectService.java
@@ -35,47 +35,6 @@ public class SubProjectService {
     private final ProjectRepository projectRepository;
     private final ModelRepository modelRepository;
 
-    @Transactional(readOnly = true)
-    public SubProjectResponseDto getSubProject(Long subProjectId){
-        //TODO 추후에 메서드 분리 필요
-        List<AnnotationHistory> histories = annotationHistoryRepository
-                .findAllBySubProjectId(subProjectId)
-                .stream()
-                .sorted(Comparator.comparing(AnnotationHistory::getCreatedAt)) // startedAt 기준 정렬
-                .toList();
-
-        List<AnnotationHistorySummaryDto> historyDtos = IntStream.range(0, histories.size())
-                .mapToObj(i -> {
-                    AnnotationHistory h = histories.get(i);
-                    return new AnnotationHistorySummaryDto(
-                            h.getId(),
-                            i+1,
-                            h.getCreatedAt(),
-                            h.getCompletedAt()// 1번부터 시작
-                    );
-                })
-                .toList();
-
-        Long latestAnnotationHistoryId = histories.isEmpty()
-                ? null
-                : histories.get(histories.size() - 1).getId();
-
-
-        //TODO 나중에 실제 userId로 변경 필요
-        Long userId =  1L;
-
-        Project project = getProjectBySubProjectId(subProjectId, userId);
-
-        List<ModelSummaryDto> modelDtos = getProjectModels(project.getId());
-
-        return new SubProjectResponseDto(
-                subProjectId,
-                historyDtos,
-                latestAnnotationHistoryId,
-                modelDtos,
-                project.getModelType()
-        );
-    }
 
     @Transactional
     public void markTilingAsComplete(Long subProjectId){
@@ -83,29 +42,5 @@ public class SubProjectService {
                 .orElseThrow(() -> new IllegalArgumentException("SubProject not found: " + subProjectId));
 
         subProject.markTilingCompleted();
-    }
-
-    private List<ModelSummaryDto> getProjectModels(Long projectId){
-        List<ProjectModel> models = projectModelRepository.findByProjectIdOrderByCreatedAt(projectId);
-
-        return models.stream()
-                .map(projectModel -> {
-                    Model model = projectModel.getModel();
-                    return new ModelSummaryDto(model.getId(), model.getName());
-                })
-                .toList();
-    }
-
-    private Project getProjectBySubProjectId(Long subProjectId, Long userId) {
-        SubProject subProject = subProjectRepository.findById(subProjectId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.SUB_PROJECT_NOT_FOUND));
-
-        Project project = projectRepository.findProjectWithUserBySubProjectId(subProjectId);
-
-        if (!project.getUser().getId().equals(userId)) {
-            throw new BusinessException(ErrorCode.NO_PROJECT_ACCESS);
-        }
-
-        return project;
     }
 }
